+++ internal/provider/provider.go
@@ -119,6 +119,7 @@ func (p *GigahostProvider) Resources(_ context.Context) []func() resource.Resour
 		NewDNSRedirectResource,
 		NewSSHKeyResource,
 		NewServerResource,
+		NewServerIPReverseResource,
 	}
 }